delete jaar_refer and redistribute str functions to graph centric files
