--- src/f01_road/jaar_refer.py
@@ -1,70 +1,0 @@
-def green_str():
-    return "green"
-
-
-def LightSeaGreen_str():
-    return "LightSeaGreen"
-
-
-def blue_str():
-    return "blue"
-
-
-def purple_str():
-    return "purple"
-
-
-def red_str():
-    return "red"
-
-
-def darkred_str():
-    return "darkred"
-
-
-def black_str():
-    return "black"
-
-
-def bob_str() -> str:
-    return "Bob"
-
-
-def buz_str() -> str:
-    return "Buz"
-
-
-def car_str() -> str:
-    return "Carl"
-
-
-def joc_str() -> str:
-    return "Joc"
-
-
-def luc_str() -> str:
-    return "Luca"
-
-
-def ell_str() -> str:
-    return "Ellery"
-
-
-def ric_str() -> str:
-    return "Rico"
-
-
-def sue_str() -> str:
-    return "Sue"
-
-
-def xio_str() -> str:
-    return "Xio"
-
-
-def yao_str() -> str:
-    return "Yao"
-
-
-def zia_str() -> str:
-    return "Zia"
