+++ src/f07_keep/keep_graphic.py
@@ -1,26 +1,78 @@
 from src.f00_instrument.plotly_toolbox import conditional_fig_show
-from src.f01_road.jaar_refer import (
-    LightSeaGreen_str,
-    green_str,
-    purple_str,
-    blue_str,
-    red_str,
-    darkred_str,
-    black_str,
-    bob_str,
-    buz_str,
-    car_str,
-    joc_str,
-    luc_str,
-    ell_str,
-    ric_str,
-    sue_str,
-    xio_str,
-    yao_str,
-    zia_str,
-)
 from src.f01_road.finance import default_money_magnitude as default_point_magnitude
 from plotly.graph_objects import Figure as plotly_Figure, Scatter as plotly_Scatter
+
+
+def green_str():
+    return "green"
+
+
+def LightSeaGreen_str():
+    return "LightSeaGreen"
+
+
+def blue_str():
+    return "blue"
+
+
+def purple_str():
+    return "purple"
+
+
+def red_str():
+    return "red"
+
+
+def darkred_str():
+    return "darkred"
+
+
+def black_str():
+    return "black"
+
+
+def bob_str() -> str:
+    return "Bob"
+
+
+def buz_str() -> str:
+    return "Buz"
+
+
+def car_str() -> str:
+    return "Carl"
+
+
+def joc_str() -> str:
+    return "Joc"
+
+
+def luc_str() -> str:
+    return "Luca"
+
+
+def ell_str() -> str:
+    return "Ellery"
+
+
+def ric_str() -> str:
+    return "Rico"
+
+
+def sue_str() -> str:
+    return "Sue"
+
+
+def xio_str() -> str:
+    return "Xio"
+
+
+def yao_str() -> str:
+    return "Yao"
+
+
+def zia_str() -> str:
+    return "Zia"
 
 
 def add_river_rect(
